backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.3.4
beanie==1.26.0
motor==3.6.1
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
aiofiles==24.1.0
python-dotenv==1.0.1

# AI and integrations
google-generativeai==0.5.2
httpx==0.27.0
